Unify TestStream and TestBatch SQL connectors
